Get rid of separate `SubscriptionItemSubject`
