+++ testutil-server/src/main/java/io/spine/testing/server/SubscriptionActivator.java
@@ -50,7 +50,7 @@ public final class SubscriptionActivator implements StreamObserver<Subscription>
         subscriptionService.activate(subscription, updateObserver);
     }
 
-    @SuppressWarnings("AvoidThrowingRawExceptionTypes") // The real type is unknown.
+    @SuppressWarnings("AvoidThrowingRawExceptionTypes") // The real type is unknown at compile time.
     @Override
     public void onError(Throwable t) {
         throw new RuntimeException(t);
